requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.1
dj-database-url==2.1.0
Django==5.0.3
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
django-mptt==0.16.0
freezegun==1.4.0
idna==3.6
pillow==10.2.0
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
